Configures Travis CI
Makes lots of changes to get our tests running in Chromium 37
which is what is available in Travis. Some of them probably
can't hurt in general.

The biggest change is that extended components are now auto
wrapped. This fixes race conditions where they rely on the
main binary and its polyfills being already loaded.