Avoid emoji in example code
Its an interesting idea to have an emoji in the code. Most editors,
however, will show the char as a an empty square. Also the code
examples on Github and the default view of .js files on Github
will show a replacement char.

All in all it creates unnecessary confusion when new people try to
understand what AMP is all about.