Adds some comments and refactors some of the presubmit-check for clarity
Some of the functions had missing or incorrect comments so this change adds
some more comments for consistency and makes a change to use a vinyl file
where it was specified in the comments but not used yet in the code. Also,
I renamed one of the functions to better reflect what was being tested (any vs all)
as it didn't make sense based on the current logic.