Fixes issue with build generating undefined for env var
Babel is replacing instances of process.env.NODE_ENV with undefined
for the non-production build [gulp build] since we aren't defining the
value. This results in confusing output in the amp.js artifact.
This change causes the undefined to be replaced by "development" by specifying
the NODE_ENV value as part of the gulp build target.