Initial contribution of the validator written by @gregable and @powdercloud of Google, with credits given to original versions for the css parser/tokenizer and the html parser in the file headers.
This is the validator which is already used by the amphtml developer mode;
see also https://github.com/ampproject/amphtml#the-amp-validator

README.md describes how to build this; there's currently only a minimal
ad-hoc build script, build.py, which should work for Ubuntu 14.

We still need to contribute our tests (we do have tests), so this is
only a first step.