Open source the feature tests for the AMP Validator.
This is still not integrated with the rest of the AMP build, so it's
piling onto build.py for now, taking the shortest route to open source
this rather than do a lot of internal refactoring first
(to be fair even this took some refactoring ...).

validator.js: Contains the logic for the test. This uses jasmin and
assert.js, the library that comes with NodeJS. I think this library is
fairly weak but for this particular purpose it was the easiest to use
because it keeps our node dependencies inside Google smaller and the
main usage here is to compare with the output of golden files.  On a
positive note, validator_test.js is run through the closure compiler so
we have that going for us. The diff in build.py makes it fairly easy to
see what's going on:

* CompileValidatorTestMinified runs closure to build validator_test.js
  with its transitive closure of dependencies, and the actual test
  method exported.

* GenerateValidatorTest refines this into a script with #! line for
  nodejs at the top, and a relatively trivial snippet of Javascript to
  invoke the test, with the testdata dir as argument.

* RunValidatorTest runs the test.

The tests in testdata/feature_tests are structured so that the .html
file contains descriptive comments explaining what the test is about,
and the .out files contain the validator output for the respective test.