Open-source additional files that are part of the AMP Validator by @Gregable and @powdercloud.
css-selectors.js: CSS selector parser (probably soon to be used)
htmlparser_test.js: Test for htmlparser.js
parse-css_test.js: Test for parse-css.js, tokenize-css.js, and
css-selectors.js.
json-testutil.js: Helper library for parse-css_test.js.

Updates to build.py to make it so that all validator tests are
executed. The approach is that the tests are compiled with
closure - each of them separately, and then there's GenerateTestRunner
which makes an executable that imports ("requires") the compiled
("minified") tests. For simplicity I've lumped them together for the
purposes of this runner. On a positive note, this new approach
no longer requires putting the entire test into a function, so it
should read a bit more idiomatic.