Slight refactoring of the layout logic.
No behavior change in practice is intended. I've extracted out a
function, determineLayout, which *only* figures out the proper value
of the layout enum for a given element, without other side
effects. This makes it easier (for me) to understand which layout
applies. This comes at the cost of a few more queries
(hasNaturalDimensions, getAttribute), hope those aren't so bad.

The other tweaks are minor and perhaps more can be done later,
but basically in applyLayout_, the layout gets nailed to a const
at the top, which separates that out.

I noticed that the code does layoutAttr.trim(), and I'd like to trim
that off while I'm at it for simplicity. Should be safe since the
validator has only allowed exact matches thus far.

@dvoytenko Thanks a lot for helping me understand this today.