Minor fixes to the templates component of the validator. These seem to be working without the fix, but let's get this cleaned up.
For some reason, calling parsedSpec.valueHasUnescapedTemplateSyntax seems to
work, even though parsedSpec is a ParsedAttrSpec and
valueHasUnescapedTemplateSyntax is only defined on ParsedTagSpec. Our only
theory is that closure has a bug which is making this function a global, since
it doesn't access any member variables. This only works in some contexts,
though.