Update the URL when a fragment link to that navigates the current document is clicked.
Also fixes fragment navigation in Firefox. It does that by deleting the code path that chose `document.body`. At least with Safari 9 all other major browsers appear to be supporting `document.scrollingElement`, to that it should be safe to pick the object here that is correct for FF. I might be missing something, though.