Improve GPT support and add Glade as an experiment
Category exclusions in GPT are properly supported and can be either a
string or an array of strings.

In the documentation and the tests, the TFCD attribute is properly
passed as a number instead of a boolean.

Glade, a new tag that is lighter and faster than GPT, is added as an
experiment on 1% of the AMP DoubleClick requests.