Add contentsWhitelist.
If there's a match for the contents of the file for this regex,
then the particular part of the presubmit is disabled. So,
if someone were to put require('goog.string.startsWith'); at the
top of the file, we assume that occurrences of 'startsWith'
within this file are OK. And same for 'endsWith'.

PS: I haven't tested this, so consider it a sketch if this works
as is it's an accident.