Remove hiding of out of viewport ads.
This has proven to not work. In WebKits the iframe itself still renders, just with 0 width dimensions which is bad in many ways. It may still be good to hide the body inside of the ad, but that is for a different PR.

While making this change I discovered that a merge error disabled many tests for ads. This PR fixes them all and removes the intersection tests which we are planning to redo.