Fixes for amp-soundcloud.
The check for the visual attribute seems incorrect. It says
if (visual) ..., but if visual were "false", then this would
evaluate to true, and so the visual mode would be enabled.
So I'd like to compare with "true" here, and I've edited the
spec to state that if set the value must be "true" or "false".
Also clarified the spec a bit more with respect to required / optional
attrs.

PS: I haven't run the tests yet (my nodejs is too old it seems),
will rely on our Github setup to do that for convenience. So please
stand by til the check is done.