Use amp.validator.ValidationError.Code for CSS errors.
This makes it so that the errors generated by the CSS parser
easily transfer to the AMP Validator as error codes and
params. Therefore, instead of the generic CSS_SYNTAX error
which takes a string parameter in its format string for the
actual error message, we now have several error codes which
can be translated into different languages. All these new
error codes map to the AUTHOR_STYLESHEET_PROBLEM error category.