2nd attempt for fixing <link rel="mask-icon" ...>
First attempt missed the fact that this also has a non-standard color attribute
associated with it. That color attribute doesn't make sense in context of a
multi-valued rel, so moved this to it's own tagspec. That'll improve errors.
Since this is new and won't be cache rewritten, I thought it would
be a good case for turning on https-only no-relative URLs. Can always relax
later.

Also noticed that the sorting order got a little out of whack for the main link
rel= lists, so cleaned that up, but should be a no-op.

Github #2047