Add tagname to STYLESHEET_TOO_LONG format string.
This is so that the format string can be translated without
translating the tagname. It makes it consistent with the other
tags.

To get this done, I renamed the existing enum entry to
STYLESHEET_TOO_LONG_OLD_VARIANT, and introduced a new entry.
The new one is the one we'll using going forward.