Simplify extension code generation.
If you have a pending pull request with a new extension you may need to adapt to this.

With this CL we no longer copy extensions and insert their CSS. Instead they just depend on the CSS if any via a generated JS file.

The main benefit of this is that the main source of an extension no longer lives in a generated file which unbreaks source maps in production for this important use case.

Fixes #2170