Add URL validation for URLs in CSS stylesheets. This distinguishes the font urls and the image urls. For both cases, we can specify a set of allowed protocols, and whether they're allowed to be relative and empty.
For starters, protocols are 'http', 'https', and they're allowed
to be relative and empty.

Emptiness is a new thing, this is because quite a few URLs
in user stylesheets on the web are specified as url(''). This makes
no sense but is probably someone trying to make the minimal edits
to get a stylesheet to pass our checker. So for now, we allow it,
but it is easy to disallow it later.