amp-accordion validation
This requires introducing child tag validation, as in, for a given
tag, limit which tags names are allowed for children.

TagSpec::child_tags is a field containing a ChildTagSpec, which specifies the child tags that are allowed. Included in this change:
- number of child tags (exact match)
- possible names for first child tags
- possible names for all child tags

The matching is done by name, to make it so that even if the child tags don't validate for some other reason, we can unambiguously refer to them and check them and generate error messages. If we were to match by TagSpecName, then we could only refer to them unambiguously if they matched a specific TagSpec themselves.

The ChildTagMatcher is a new matcher which is put onto the TagStack, previously called TagNameStack. This makes sense because we need scoping for each tag in order to count the children.

I intentionally kept the constraints that are checkable simple and relatively specific to amp-accordion, but ChildTagMatcher can obviously be extended with different checks later.