Implement AMP-side click measurement. (#2969)
Introduces a redirect-avoidance mechanism that enables referrers to track a link invocation without an intermediate redirect by passing a URL to AMP via a fragment param. This change is experiment guarded.

Also introduces a small new binary for use on the referrer side (e.g. inside of an ad creative) that helps with constructing the respectives URLs and preconnects to AMP.

Primary implementation of #2934