A NodeJS API for the validator.
The implementation is in the already existing index.js.
By default, uses the latest deployed validator script from
https://cdn.ampproject.org/v0/validator.js, but also
accepts an argument for some other validator (usually a filename,
e.g., 'dist/validator_minified.js').

The unfortunate reality of Javascript is that this requires
some async programming, so I've tried to learn from
http://www.html5rocks.com/en/tutorials/es6/promises/ how to do
this with promises. But please look for anti-patterns, especially
at the API boundary (like, the way the exported validateString
function works).

I've put a couple classes for the exported types ValidationResult,
ValidationError into the module and documented them there. This
structure may not change in the future, but minor things under the
hood could.

index_test.js contains a simple unittest for this API.

I've updated README.md with documentation.