More polymer elements and automatic resizing for (#3123)
the validator webui.

I refactored this a bit to make it less monolithic. More work
remains to be done.

In addition, using the iron-resizable abstraction to make the
editor resize as the viewport size changes, and fill it
to a reasonable extent. Also, the list of errors near the bottom
of the screen is now always scrollable. This makes the overall
experience a bit smoother.