Validation for amp-kaltura-player. Similar to amp-youtube etc., in this (#3181)
case the data-partner attribute is mandatory and everything else
is optional. I would be happy to validate more, e.g., with regexes,
but don't know what the specs are for these parameters.

Also one small fix in amp-kaltura-player.js itself, where the error
message differs from the attribute that's being checked.
Also a small edit to the amp-kaltura-player.md file to indicate that
data-partner is a mandatory attribute.