Turn on more aggressive compiler optimizations.
- allows devirtualization of private methods.
- allows renaming of private properties and methods.
- Aliases `.prototype.` pattern where useful.
- other less important optimizations.

Implements a coding convention for AMP that marks all properties not ending in `_` as exported (so they can travel between compilation units).

This change by itself has little size impact, but should make our code quite a bit tighter to execute, because many virtual private method calls because function calls. Primarily this is a preparation to turn on private property renaming when the upstream change lands in closure compiler.