A NodeJS API for the validator. (#3086)
* A NodeJS API for the validator.

The implementation is in the already existing index.js.
By default, uses the latest deployed validator script from
https://cdn.ampproject.org/v0/validator.js, but also
accepts an argument for some other validator (usually a filename,
e.g., 'dist/validator_minified.js').

The unfortunate reality of Javascript is that this requires
some async programming, so I've tried to learn from
http://www.html5rocks.com/en/tutorials/es6/promises/ how to do
this with promises. But please look for anti-patterns, especially
at the API boundary (like, the way the exported validateString
function works).

I've put a couple classes for the exported types ValidationResult,
ValidationError into the module and documented them there. This
structure may not change in the future, but minor things under the
hood could.

index_test.js contains a simple unittest for this API.

I've updated README.md with documentation.

* Make index.html completely static unless a different validator.js
is injected (now by replacing the validator.js string directly).
Call the key in the json Contents, not contents. This makes it
simpler for Go, because Go likes to upper-case its field names.

* - Name this pacakge 'amp-validator'.
- Fix param name to opt_validatorJS and fix type decl for promise.

* Be specific about the version I tested with.
Add a note that we'll try to port to earlier versions if
there's sufficient interest.

* Add comment about sandbox.

* Don't add ValidationResult / ValidationError to the exports.

* Introduce a getInstance function which returns the promise.
Make validateString a method which directly returns the validation result.
Updated test and README.md.

* fix Promse type decls

* add clarifying comment to readFromStdin

* move readFromReadable to then, add comment about resolve

* use chaining

* more chaining

* use Promise.resolve

* explictly map back to the string keys from the enum values

* clang format

* for now do not invert the enum maps

* make the exceptions carry a useful message from the get go

* remove double reject