In the validator UI allow URLs without protocol.
Tweak the regex a bit to get this done.
When the protocol is left off, change the input so the user
can see what's happening.

We should consider doing this in Javascript if we want to make
this a lot more elaborate, to avoid getting stuck in this rabbit hole:
https://mathiasbynens.be/demo/url-regex :-)