Pass scrollHeight and scrollWidth to documentLoaded event
It made no sense to pass the viewport’s width and the document’s
scrollHeight. We should either pass the viewport’s width and height, or
the document’s scrollWidth and scrollHeight.

Since you can determine the viewport’s size from the viewer’s iframe, I
chose to send the scrollWidth and scrollHeight.