Set Owner on Resource instance
Imagine an amp-element (A) is wrapped inside some other element (B)
inside a containing amp-carousel.  If A already checked to see if it's
owned, it will have a cached `Resource#owner`. If you then set ownership
of B (as amp-carousel currently does), A will only use it's cached
ownership and will never know that is now owned (implicitly) through B.

- [ ] Needs tests