Set the Node.js version to 0.10.25, the Node.js version from Ubuntu Trusty. (#3764)
I have verified that index.js runs OK with that.
In index_test.js, fix an es6 construct to be old-style. But even with that
it won't pass with 0.10.25.
There's one remaining problem with running on Trusty: The shebang line
wants node, but on Trusty it's 'nodejs'. I will attempt to fix this in a
later revision.