Make our compilation O(extensions) instead of O(extensions**2)
Closure compiler parses all files in the glob for module ids and dependencies only to drop 99% on a typical compilation.

This change make extensions only glob themselves and not other extensions and additonally makes the glob area sparser (e.g. only `integration.js` globs the 3p code like it should be).

Significant additional work can be done to further reduce the files included in a given compilation, but this should be a good first step.