Translate element coords into viewport coords
I made a mistake in Fixes a mistake in #4826. The “viewport” we report
moves with the scroll position, while a real IntersectionObserver does
not. And, our elements didn’t move with the scroll position, while the
real one did.

So, this translates elements into their position relative to scroll,
and removes the scroll relative position from viewport.

Fixes #5090.
/cc @jasti