Attempt to move execution of gulp tasks into script.
This makes it easier to skip tasks that are not relevant to the PR.
The logic I made is pretty ad hoc thus far, but I figure that it's
a starting point for making this more sophisticated later - and
even more fine grained. But for this change I'm already hoping that
we no longer run the validator tests for runtime PRs and vice versa.

This change isn't tested except for fumbling a bit with the node
interpreter, hence it's in rough shape. Not sure how to best do this
please advise.