Fix silent Shadow DOM failure
Fixes silent Shadow DOM failure for compiled builds only in Safari. Both AMP PWA and Polymer PWA exhibit this bug.

`this` was incorrectly set to the window object within ShadowCSS, which resulted in all self-referential functions inside ShadowCSS to fail. My guess is that Closure's DCE simply stripped all of these method calls, which is why the failure was silent.

I'll survey the rest of the usages of `webcomponentsjs` within the AMP src to see if this is happening elsewhere and follow up with the `webcomponentsjs` team to see if they can switch to a less error-prone way of exporting objects.