Attempt to move execution of gulp tasks into script. (#5157)
* Attempt to move execution of gulp tasks into script.

This makes it easier to skip tasks that are not relevant to the PR.
The logic I made is pretty ad hoc thus far, but I figure that it's
a starting point for making this more sophisticated later - and
even more fine grained. But for this change I'm already hoping that
we no longer run the validator tests for runtime PRs and vice versa.

This change isn't tested except for fumbling a bit with the node
interpreter, hence it's in rough shape. Not sure how to best do this
please advise.

* add TRAVIS_COMMIT_RANGE to command

* use synchronous execution in this script

* more comments

* do not require child_process

* one more try

* print stdout

* use spawn

* add OK to console.log and console.error to get past presubmit

* do not run gulp dist for validator targets

* process the return value of spawn

* add a failing test to validator_test.js

* introduce test that passes to validator_test.js

* Improve messages in pr-check and add failing runtime test.

* remove temporary tests now that we know pr-check works

* add missing closing curlies

* * hasPrefix/Suffix -> whitelisted {starts,ends}With + whitelist
* remove console /*OK*/ -> console + whitelist
* small bugfix in isValidatorFile (missing param for hasSuffix)
* move npm run ava etc. into runtime only portion