RFC: Separate the load phase of AMP into multiple chunks. (#5536)
The general idea is that AMP can be instructed to not block the UI thread during load phase for long if the document itself is not visible.

The change may be quite impactful. It does split up work into neat chunks (about 10-20ms on desktop, longer on mobile) instead of one 400ms+ block. I validated that the number of style recalc doesn't change when the experiment is off.

- Current doesn't change `amp-shadow`. That will come in a follow up CL, but might not make as much sense.
- With the experiment not active the only thing this should change is that the initialization now goes through a bunch of micro tasks. This should behave as before in browsers with native promises.