Introduce timeout for font stylsheets in AMP. (#5911)
All external stylesheets in AMP have the sole purpose of loading fonts. While browsers have a timeout for fonts (3s or lower), they don't have a similar timeout for stylesheets. Thus a user will often wait a long time for the stylesheet, only to run into the next timeout for the font itself.

With this change, if the page has not reached ready state 1s after the response started arriving on the client, AMP will look for external stylesheets referenced in the doc and replace them with dynamically inserted stylesheets, which do not block displaying the page.

As far as I know there is no way to tell whether a stylesheet has loaded. For this reason, we use `document.readyState` as a signal, since it is blocked by external stylesheets.

This change may have significant impact on AMP's 99%ile load performance.