Cache SW: Do not cache experiments.js (#5893)
* Cache SW: Do not cache experiments.js

This explicitly disallows `v0/experiments.js` from being cached by our
SW. This is necessary because we do not host an RTV version of this
file, which is required for anything that's cached. Plus why would we
cache the experiments javascript?!

I've opted to de-duplicate a few areas into a single regex. Let me know
if this is too complicated. This will allow us to easily support v1, v2,
etc versions later on, too.

* Tests should use pathnames