Fix HTML Parser when it encounters unclosed comment at EOF. (#6182)
* Add a postinstall script for Windows.

The command line is executed both for cmd.exe and sh, the Unix shell,
depending on where it runs. Note that cmd.exe is *really* different,
but there is a small subset that's common. So the postinstall command
line works on both.

1) For Windows: /bin/sh doesn't exist, so this is an error, which we
redirect to NUL (which is the Windows version of /dev/null). Then we
try to delete NUL, which won't work and give a non-zero exit status.
But up to here it's all quiet. Now we get to
|| node postinstall-windows.js, and this just runs the windows specific
postinstall script. This script locates the Windows shim and modifies
it so that it works. The shim that we put there just calls node with
the name of the script file (index.sh), so we're skipping the index.sh
dispatch script that we use for Debian-based (where node is called nodejs)
and other Linux distros (where it's called node).

2) For Unix-like platforms, /bin/sh \"exit 0\" will have exit status 0.
NUL will be a temporary file, which we then remove. This succeeds,
so we won't run the postinstall-windows.js script.

The index.sh script makes amphtml-validator work for both Ubuntu/Debian/etc.
and other installations, but this part isn't new.

* Fix HTML Parser when it encounters unclosed comment at EOF.