Sentinel passing refactored, don't pass on element.
Previously, the sentinel had been passed both on the hash of the
iframe src, and as a data attribute. The data attribute was how
the sentinel was referenced on the runtime side, even though there
was no reason it could not simply be passed.

Now, the sentinel is only passed to the iframe as a member of an
object that is stringified and attached to the iframe name.

On the runtime side, the sentinel is now passed and expected as
a parameter to multiple functions that used to ultimately rely
on it being available as a data attribute of the iframe.