Generate fallback clientId for Foreign Fetch
We use `clientId`s to distinguish between JS context windows, ensuring
that we serve one uniform AMP JS version to all requests from the
client.

Foreign Fetch, unfortunately, does not provide a `clientId`, so we must
generate our own.

This lumps all requests from a referrer into 60 second "batches",
generating a single `clientId` for those requests. After 60 seconds from
the first request, the batch's timer is invalidated and the next
request will generate a new `clientId` from the same referrer.

60 seconds was arbitrarily chosen to prevent version lock happening when
navigating back to the same page. We could have just used the referrer
as the `clientId`, but as long as the SW remains active, there would be
no way to distinguish between a navigation back. As we want to serve the
newer JS when possible, that seemed bad.