Cache SW: Serve version with most responses
I want to avoid the situation where
`/v0/some-rarely-used-components.js` serves up a very old RTV, when we
have multiple files of the newer version that could be served.

This heavily weights the main binary, places a small weight on the
first requested file, and treats all other responses as normal. The
version with the max score is what we’ll serve.