Do not extend backoff reporting useless errors
These include already reported errors (which somehow got thrown again),
cancelled loading messages, and now a large majority of loading errors.

If we had a whole bunch of loading errors, and then a real honest to god
bug, we'd almost never receive the bug because the "reported" loading
errors would delay sending the message for so long.