force dom reflow after appending overflow
iOS9 appears to have a race condition and sometimes shows the overflow
on top of the ads, even though ALL containers and iframe appear to have
the right height set .  Adding this to see if it fixes the issue.  This
doesn’t appear to be a problem for iOS10 or Android.