Ensure that viewer handshake always happens quickly.
This fixes #6728. THe problem was that chunking can delay the handshake with the viewer for a very long time. During this time the viewer cannot tell the doc that it is now visible–and visibility triggers chunking to go into fast mode. With this change the aggressive chunking mode is only triggered when the doc has gotten into a state where it can communicate with the viewer; and the viewer communication code itself is never delayed in a chunk.