Integrate Google AdSense for Search and Adsense for Shopping (on the Custom Search Ads protocol) with AMP (#6518)
* Initial commit for CSA AMP ad integration

* update ads example

* assign throwaway variable so compiler doesn't complain

* fix lint errors

* define _googCsa in backfill function

* Add link to CSA docs in AdSense docs

* fix test error

* fix type error

* feedback from jeff

* update css example

* create overflow on resize denied if it doesn't exist

* fix from jeff's comments

* initial stab at unit tests

* first stab at unit tests

* add test for backfill

* lint issues

* create helper functions and move to bottom

* hongfei's feedback

* use Object.assign and tryparsejson

* update tests

* code review changes

change some variable names, mostly condense code

* change global.innerHeight back to initial height

global.innerHeight doesn't work because I need the height of the AMP
container at the time of requesting the ads before the iframe is
resized. Different browsers have different behaviors (some bust the
iframe, some don't). Changing to global.innerHeight broke the overflow
implementation on iOS.

* force dom reflow after appending overflow

iOS9 appears to have a race condition and sometimes shows the overflow
on top of the ads, even though ALL containers and iframe appear to have
the right height set .  Adding this to see if it fixes the issue.  This
doesn’t appear to be a problem for iOS10 or Android.

* Revert "force dom reflow after appending overflow"

This reverts commit 2040f76c5f501ad9b0e7b777d91aea2ace6ecd02.

* hide overflow of CSA iframe

I’m adding this because on ios9, the ad content is visible underneath
the overflow element, even though the height of the container and
iframe are being set correctly.

* Set iframe width to 100% to prevent ios9 from breaking the overflow implementation

* move code into functions

* Move resize handlers to own functions, register in CSA function, not resizeIframe

other minor refactoring as well

* nitpicks

* add forgotten breaks to switch statement

* Create two callback functions, one for backfill and one without.

* Merge remote-tracking branch 'ampproject/master' into csa-integration

# Conflicts:
#	ads/google/csa.js

* Inline backfill ads function

* Merge with adsense.md accidentally replaced HC link

* travis errors

* more travis errors