Throw an error if a required attribute is missing.
Implement better error checking for CSA ad integration. Currently, if
the tag is missing a necessary attribute a request still gets made.
This change ensure that they are including the right attributes for the
correct product.  We can’t do this earlier in the code because we
aren’t sure which product to call until this point.