Create a pull request template for the project.
The template should help make pull requests easier to read and overall more uniform.

- Does not propose complicated syntax for titles (or rather forbids it).
- Encourages referencing issues.

This is a bit of an RFC. Feedback is welcome. One thing I don't actually know (and the GH docs don't say), is how this template influences auto-generated PR text. It would be really annoying if that no longer happened.