Reduce URL entropy for preconnect polyfill. (#7227)
A single AMP page would already only connect to a given origin once ever 180 seconds. With this change the URL for doing this will not change within that time interval. That means that all AMP pages that run on computers with roughly correct clocks will request the same URL within a 180 seconds interval, which should make it easier for CDNs to respond efficiently.