Fix undefined reportError (#7588)
* Fix undefined reportError

Fixes a bug from #7363. Because the extensions never see a call to
`setReportError`, in their own bundles `reportError` is probably
`undefined`.

This fix makes `reportError` a global variable, so we get the same
requirements (entry-points must call `setReportError`). But now, the
extension bundles will use that global `reportError`.

* Treat setReportError like initLogConstructor

`setReportError` MUST be called from every entry point.

* Use simple console logs for 3p frame error reporting

Using the standard error reporter requires a tangle of dependencies that
are not available to 3p frames. After #5527, we can use `reportError`
like normal.

* Add comments