Support LCD amp-ad in multi-size
This check makes it only possible to do a GCD size for the primary
unit. Without the ability to put a creative wrapper on AdX we cannot
resize the container back down to 300x250.
By removing this requirement we can set the default size to 300x250 for
AdX support and allow the amp-ad container to resize up for direct sold
sponsorship larger ad sizes. The container will only resize downward
when BTF so there is no negative effect on user experience that I've
seen.
Perhaps a better way to solve this is to include a new attribute in the
amp-ad tag that allows for this option as an override. Thanks Sam
smansour@hearst

http://localhost:8000/examples/HDM-Superhero.max.html