Limit the number of elements to build during prerender (#8364)
This significantly reduces the amount of CPU needed for very large documents during prerender. The count of elements to build was chosen arbitray as "big enough that it typically should be enough to render stuff in the first viewport".

This seems to work quite well and it suggests we do have quite the opportunity to optimize things with a chunking scheduler.

The document I was testing with (http://www.refinery29.com/amp/best-pizza-nyc) does get much better with this, although it doesn't get into good shape, since it is simply too big for a browser (let alone a mobile browser) to reasonably render.