Remove timer dependency from event-helpers (listen, listenOnce) (#8472)
* Remove opt_timeout from loadPromise

It’s unused

* Allow caller to call internal unlisten

This inverts the responsibility of calling the unlistener. Now, the
caller can be given access the `unlisten` by passing a closure. The
closure is called with `unlisten` as the param, giving the caller
access. When the caller determines it is time to stop listening (due to
timeout, etc), it can manually cleanup.

* lint

* fix merge conflict

* fix test

* Fix tests