Distinguish between element-provided and extension provided services
Extension-provided services are automatically registered by the
extension (eg, CID). Element-provided services are registered by an some
configuration element (not the extension script).

For extension-provided scripts, we can depend on a service promise if
the extension is registered (amp-analytics being registered guarantees
CID).

But for element-provided, we cannot depend on the service promise just
because the extension is included. Instead, we have to wait for the
`<body>` to be ready, and see if it's there (not guaranteed).