Remove dependency on WebCrypto for origin cookie generation.
SubtleCrypto is going away for insecure origins. But we also don't really need it when not serving from a proxy. Instead produce a direct string from the entropy when that entropy was gathered with `getRandomValues`.

See https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/ZD3NWqkk-bo for context.