Update Promise PJs
Fixes https://github.com/ampproject/amphtml/issues/10730.

A very interesting bug where promise adoption (an inner promise being
returned in side a `#then` block causes the outer to adopt the inner)
would not update the state of the outer promise if the inner was still
pending.

This only appeared when:

1. The outer promise was stored, and not used as a larger chain.
2. The inner promise was pending.
3. The outer promise has a `#then` block _after_ the inner promise
  resolved.