Fix amp-web-push in unsupported environment
When amp-web-push starts up, it checks whether the browser environment
supports web push notifications. If the environment doesn't support web
push notifications, the code incorrectly returns ... into the next
method and just continues running. Eventually, configuration variables
that were supposed to be set in the initialization method we just quit
out of, are accessed, but are null, and the whole thing errors.

The web push service now exits immediately when starting to run if the
environment is unsupported. This will cause any web push widgets for
subscription or unsubscription to remain hidden, since they're not shown
until the service determines the user's subscription state, and the
service is now stopped.