Intercept XHR for trusted capable viewers
Some trusted custom AMP Viewers require the runtime to delegate all XHRs
triggered from the AMP document to the viewer to handle.

Instead of sending the XHRs directly from the AMP document, the AMP
runtime delegates the task of sending the XHR to the viewer and waits
for the viewer to return the XHR response.

This includes a polyfill for FormData#entries, as the XHR service needs
to be able to read the entries in FormData if it is passed to `body` in
order to make it a strucured cloneable.