Disable widget when subscribing or unsubscribing
Until a subscription is complete or canceled, or until an unsubscription
is complete, disable the widget to prevent multiple taps, and then
enable the widget after the task is complete.
This commit changes the logging from development logging, which is
removed from the final build, to user logging, which is retained in the
final build. This should be helpful for debugging production sites.