Initialize AmpDocService in Xhr service in tests
This is a pre-requitesite for #11294.

Currently, AmpDocService is only initialized if not in test mode.
However, the about-to-be-added XHR interceptor feature also needs to
access AmpDocSerivce, and that feature requires extensive functional
testing. If AmpDocService is only initialized outside of test mode, the
XHR interceptor logic becomes untestable.

To prepare for testing XHR interceptor, AmpDocService needs to be always
initialized regardless of the mode (may initialize to null if doc is not
in single doc mode). Instead, the check for test mode is moved to the
call site.