Polyfill FormData#entries
This is a pre-requitesite for #11294.

Formdata#entries is not supported by IE, Edge and Safari, but is useful
for accessing content of the a `FormData` object. On IE, Edge and
Safari, the only available method on `FormData` is `append`, which means
that client code can't read from a `FormData`. Reading entries in a
`FormData` will be required by #11294 to serialize a `FormData` object
into a plain old JavaScript object so that it can be passed to
`postMessage()` (i.e., can be cloned using the structured clone
algorithm).