Better handle installation isolation
Instead of storing the window where the polyfill is being installed onto
the function object, pass the window directly to the constructor
function by using `bind`. This ensures that separate installation for
distinct windows are isolated with different constructor function, each
having a reference to the native FormData constructor for the respective
window. Sharing state on the FormDataPolyfill function object across
window is problematic as you can't install the polyfills to multiple
windows.

Instead of copying the prototype to implicitly satisfy the precondition
check that prevents re-installation of the polyfill on the same window,
explicitly check that win.FormDataNative_ has been patched to prevent
re-entry of the `install` method on the same window.

For the tests, since `env.win.FormData` will be patched to the polyfill
function after calling `install`, `env.win.FormData` won't have anything
in the prototype. Instead, the tests should assert on the native
constructor's prototype.

Misc fix: Following AMP's convention of prefixing it with `opt_` and
giving it a default value (even if it's `undefined`).