Store notif perm earlier to allow popup
Store the notification permission before the user clicks the Subscribe
button, because the permission may currently be granted, and we can
avoid opening the popup in the future if the permission is granted.

The user might have previously subscribed and then cleared their
browser data without resetting their permission, and in this case we
don't have to.

We can't check the notification permission when the user clicks
Subscribe because waiting on promises after clicking buttons triggers
the browser's popup blocker. So we have to check and store the
permission before the user clicks Subscribe and use the stored result
when the user finally clicks Subscribe.

Also store the notification permission right after the user interacts
with the prompt, so we can store the result of blocking the prompt to be
used on next page load.